backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
github3.py==4.0.1
redis==5.0.1
celery==5.3.4
pytest==7.4.3
langchain==0.0.340
python-jose==3.3.0
passlib==1.7.4 